requirements.txt
PyPDF2
groq
chromadb
sentence-transformers
nltk
fpdf
torch
streamlit
python-dotenv
faiss-cpu